Refactor Maxwell classes into ExplicitField, MaxwellField and MaxwellMultiple
