+++ src/quvac/field/external_field.py
@@ -7,8 +7,8 @@ import os
 import numpy as np
 
 from quvac.field.abc import Field
-from quvac.field.paraxial_gaussian import ParaxialGaussianAnalytic
-from quvac.field.maxwell import ParaxialGaussianMaxwell, ParaxialGaussianMaxwellMultiple
+from quvac.field.gaussian import GaussianAnalytic
+from quvac.field.maxwell import MaxwellMultiple
 
 
 class ExternalField(Field):
@@ -26,6 +26,7 @@ class ExternalField(Field):
     def __init__(self, fields_params, grid, nthreads=None):
         self.fields = []
         self.grid = grid
+        self.grid.get_k_grid()
 
         self.nthreads = nthreads if nthreads else os.cpu_count()
 
@@ -39,43 +40,23 @@ class ExternalField(Field):
         for field_params in new_params:
             self.setup_field(field_params)
 
-        # for field_params in fields_params:
-        #     self.setup_field(field_params)
-
     def setup_field(self, field_params):
         if isinstance(field_params, list):
             field_type = 'maxwell'
         else:
             field_type = field_params["field_type"]
+
         match field_type:
             case "paraxial_gaussian_analytic":
-                field = ParaxialGaussianAnalytic(field_params, self.grid)
-            case "paraxial_gaussian_maxwell":
-                field = ParaxialGaussianMaxwell(field_params, self.grid, nthreads=self.nthreads)
+                field = GaussianAnalytic(field_params, self.grid)
             case "maxwell":
-                field = ParaxialGaussianMaxwellMultiple(field_params, self.grid, nthreads=self.nthreads)
+                field = MaxwellMultiple(field_params, self.grid, nthreads=self.nthreads)
             case _:
                 raise NotImplementedError(f"We do not support '{field_type}' field type")
         self.fields.append(field)
-
-    def allocate_buf(self):
-        try:
-            for i in range(3):
-                self.E_out_[i] *= 0.
-                self.B_out_[i] *= 0.
-        except:
-            self.E_out_= [np.zeros(self.grid.grid_shape, dtype=np.complex128) for _ in range(3)]
-            self.B_out_= [np.zeros(self.grid.grid_shape, dtype=np.complex128) for _ in range(3)]
             
     def calculate_field(self, t, E_out=None, B_out=None):
-        # self.allocate_buf()
         for field in self.fields:
             E_out, B_out = field.calculate_field(t, E_out=E_out, B_out=B_out)
-        # if E_out is None:
-        #     E_out = [np.zeros(self.grid.grid_shape) for _ in range(3)]
-        #     B_out = [np.zeros(self.grid.grid_shape) for _ in range(3)]
-        # for i in range(3):
-        #     E_out[i] += np.real(self.E_out_[i])
-        #     B_out[i] += np.real(self.B_out_[i])
         return E_out, B_out
